requirements.txt
numpy>=1.20.0
Pillow>=9.0.0
PyQt5>=5.15.0
imagehash>=4.2.0
av>=9.0.0 